refactor(ui): address review — robust loadImage, clamp/sanitize hardening, extract preview hook
